refactored the new render cards component
refactored the new cards component
