--- src/pages/bookmarkpage/bookmark.scss
@@ -1,1 +1,0 @@
-@import '../../stylesheet/main.scss';
