+++ src/pages/homepage/home.scss
@@ -18,13 +18,6 @@
 			gap: 6.4rem;
 		}
 	}
-	&-recommended-section {
-		h1 {
-			margin-top: 4rem;
-			margin-bottom: 3.2rem;
-			@include font-hlg-light;
-		}
-	}
 	//to hide swiper button responsively
 	.swiper-button-next,
 	.swiper-button-prev {
